+++ src/Entity/Tile.java
@@ -11,8 +11,8 @@ public class Tile extends Entity {
 
     public int kind;
 
-    public Tile(double x, double y, int kind) {
-        super(x, y, (double) TileMap.tileSize, (double) TileMap.tileSize, (double) 0);
+    public Tile(int x, int y, int kind) {
+        super(x, y, TileMap.tileSize, TileMap.tileSize, 0);
         this.x = x;
         this.y = y;
         this.kind = kind;
@@ -23,8 +23,7 @@ public class Tile extends Entity {
         }
     }
 
-    public void draw(Graphics2D g) {
-    }
+    public void draw(Graphics2D g) {}
 
     public boolean hit(int amount) {
         return kind == TileMap.WINDOW;
