+++ src/Entity/Weapon/Remington.java
@@ -7,7 +7,7 @@ public class Remington extends Weapon {
 
     public static int maxAmmo = 12;
 
-    public Remington(double x, double y, double w, double h, double orient, int ammo) {
+    public Remington(int x, int y, int w, int h, float orient, int ammo) {
         super(x, y, w, h, orient, ammo);
         setConstants();
     }
@@ -17,7 +17,7 @@ public class Remington extends Weapon {
         type = 2;
         bVel = 7;
         amount = 6;
-        spread = 0.3;
+        spread = 0.3f;
         damage = 15;
         rate = 400;
     }
